# Samples/swagger_tvscraper.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TvScraper.Repository",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/TvScraper/GetTvShowsFromDb/page/{page}": {
      "get": {
        "tags": [
          "TvScraper"
        ],
        "operationId": "GetTvShowsFromDb",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShowsDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowsDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowsDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/TvScraper/UpdateTvShowDb": {
      "get": {
        "tags": [
          "TvScraper"
        ],
        "operationId": "UpdateTvShowDb",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/TvScraper/GetConcenatedShows": {
      "get": {
        "tags": [
          "TvScraper"
        ],
        "operationId": "GetConcenatedShows",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TvShowDb"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Actor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ShowsDTO": {
        "type": "object",
        "properties": {
          "totalShows": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "tvShows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowDb"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TvShowDb": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cast": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actor"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}